# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/schools/{id}": {
      "get": {
        "operationId": "SchoolsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Schools"]
      }
    },
    "/api/v1/admin/schools": {
      "get": {
        "operationId": "AdminSchoolsController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminSchools"]
      },
      "post": {
        "operationId": "AdminSchoolsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchoolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminSchools"]
      }
    },
    "/api/v1/admin/schools/{id}": {
      "patch": {
        "operationId": "AdminSchoolsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchoolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminSchools"]
      }
    },
    "/api/v1/admin/schools/{id}/archive": {
      "post": {
        "operationId": "AdminSchoolsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminSchools"]
      }
    },
    "/api/v1/admin/schools/{id}/unarchive": {
      "post": {
        "operationId": "AdminSchoolsController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminSchools"]
      }
    },
...
    "/api/v1/enrollments": {
      "get": {
        "operationId": "RegistrationsController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Registrations"]
      },
      "post": {
        "operationId": "RegistrationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Registrations"]
      }
    },
    "/api/v1/enrollments/guided": {
      "post": {
        "operationId": "RegistrationsController_createGuided",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidedEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Registrations"]
      }
    },
    "/api/v1/enrollments/{id}": {
      "get": {
        "operationId": "RegistrationsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Registrations"]
      }
    },
    "/api/v1/enrollments/{id}/submit": {
      "post": {
        "operationId": "RegistrationsController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Registrations"]
      }
    },
    "/api/v1/enrollments/{id}/cancel": {
      "post": {
        "operationId": "RegistrationsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Registrations"]
      }
    },
    "/api/v1/admin/enrollments/families/{familyId}/guided": {
      "post": {
        "operationId": "AdminRegistrationsController_createForFamily",
        "parameters": [
          {
            "name": "familyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidedEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminRegistrations"]
      }
    },
...
    "/api/v1/admin/enrollments/{id}": {
      "get": {
        "operationId": "AdminRegistrationsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminRegistrations"]
      }
    },
    "/api/v1/admin/enrollments/{id}/start-review": {
      "post": {
        "operationId": "AdminRegistrationsController_startReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminRegistrations"]
      }
    },
    "/api/v1/admin/enrollments/{id}/approve": {
      "post": {
        "operationId": "AdminRegistrationsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminRegistrations"]
      }
    },
    "/api/v1/admin/enrollments/{id}/reject": {
      "post": {
        "operationId": "AdminRegistrationsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminRegistrations"]
      }
    },
    "/api/v1/admin/enrollments/{id}/request-correction": {
      "post": {
        "operationId": "AdminRegistrationsController_requestCorrection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminRegistrations"]
      }
    },
...
    "/api/v1/onboarding/enrollments/guided": {
      "post": {
        "operationId": "OnboardingRegistrationsController_createGuided",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidedEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["OnboardingRegistrations"]
      }
    },
    "/api/v1/onboarding/enrollments/{id}": {
      "get": {
        "operationId": "OnboardingRegistrationsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["OnboardingRegistrations"]
      }
    },
...
  },
  "info": {
    "title": "School Transport API",
    "description": "Canonical REST contract for the school transport MVP.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "CreateSchoolDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSchoolDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateRegistrationDto": {
        "type": "object",
        "properties": {}
      },
      "GuidedEnrollmentDto": {
        "type": "object",
        "properties": {}
      },
      "RejectRegistrationDto": {
        "type": "object",
        "properties": {}
      },
      "CorrectionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}